1. 결제, 구독 페이지 헤더 변경 2. MemberController 수정 3. noImage.png 추가로 책에 대한 그림 없을때 처리하기

+++ src/main/java/com/example/bookangel/mappers/CouponMapper.java
@@ -11,7 +11,7 @@ public interface CouponMapper {
     public Long checkCoupon(CouponVO couponVO);
 
     // 쿠폰 사용
-    public long useCoupon(CouponVO couponVO);
+    public int useCoupon(CouponVO couponVO);
 
     // 기업 쿠폰 리스트 조회
     public List<CouponVO> companyCouponList(long memberNum);
